mncharsel: fold the static header into the translation unit
The unit is matched and linked, so its private data, types, and constants
belong in the .c rather than a separate header.
